requirements.txt
absl-py==1.1.0
alembic==1.8.0
arrow==1.2.2
asn1crypto==0.24.0
asttokens==2.0.5
backcall==0.2.0
# basic-fne===1.0.347-c8f0d2c5
boto3==1.12.46
botocore==1.15.49
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.6
clang==5.0
click==8.1.3
cloudpickle==2.0.0
codecarbon==2.1.0
cryptography==2.1.4
databricks-cli==0.16.8
dataclasses==0.6
decorator==5.1.1
docker==5.0.3
docutils==0.15.2
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
Flask==2.1.2
flatbuffers==1.12
future==0.18.2
fuzzywuzzy==0.18.0
GitPython==3.1.27
gunicorn==20.1.0
horovod==0.23.0
hpai-utils==0.0.21
idna==2.6
importlib-metadata==4.8.1
ipython==8.4.0
jedi==0.18.1
jmespath==0.10.0
joblib==1.1.0
keras==2.9.0
keyring==10.6.0
keyrings.alt==3.0
libclang==14.0.1
llvmlite==0.38.1
lucid==0.3.8
matplotlib-inline==0.1.3
mlflow==1.26.1
more-itertools==8.13.0
numba==0.55.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-flask-exporter==0.20.2
prompt-toolkit==3.0.30
protobuf==3.18.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pycrypto==2.6.1
Pygments==2.12.0
PyGObject==3.26.1
pynndescent==0.5.7
pynvml==11.4.1
PyOpenGL==3.1.6
pyparsing==3.0.9
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
pytz==2022.1
pyxdg==0.25
PyYAML==6.0
querystring-parser==1.2.4
requests==2.18.4
s3transfer==0.3.7
scikit-learn==1.1.1
scipy==1.8.0
SecretStorage==2.3.1
six==1.15.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
ssh-import-id==5.7
stack-data==0.3.0
tensorboard==2.9.1
tensorflow-addons==0.17.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-rocm==2.9.1
threadpoolctl==3.1.0
torch==1.12.0
tqdm==4.64.0
traitlets==5.3.0
typeguard==2.12.1
typing-extensions==3.7.4.3
umap-learn==0.5.3
unattended-upgrades==0.1
urllib3==1.22
wcwidth==0.2.5
xgboost==1.5.2
